+++ CHANGELOG.md
@@ -36,7 +36,7 @@ For details about compatibility between different releases, see the **Commitment
 
 - Increased the maximum gRPC message size to 16MB.
 - Gateways which have been deleted are now automatically disconnected by the Gateway Server.
-- Disabled devices that are registered on a different cluster.
+- Mark off and hide and restrict access to end devices that are registered on a different cluster in the Console.
 
 ### Fixed